Swap tasks Fix description of iterators task 2
