+++ iterators_task_2/student/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(key_value_iterator_HOMEWORK VERSION 1.0.0)
+project(accumulator VERSION 1.0.0)
 
 set(sources
     "src/impl.cpp"
@@ -17,15 +17,17 @@ set(LESSONS_ROOT ${ROOT_DIR}/../../)
 
 set(TEST ON)
 
-add_library(key_value_iterator_HOMEWORK SHARED ${sources})
+add_library(accumulator SHARED ${sources})
 
-set_target_properties(key_value_iterator_HOMEWORK PROPERTIES
+set_target_properties(accumulator PROPERTIES
                         C_STANDARD 11
                         CXX_STANDARD 11
                         PUBLIC_HEADER "${headers}"
 )
 
-target_include_directories(key_value_iterator_HOMEWORK
+
+
+target_include_directories(accumulator
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
