+++ iterators_task_2/student/tests/unit_tests/src/maintest.cpp
@@ -1,137 +1,157 @@
 #include "maintest.h"
+#include "impl.h"
+#include <numeric>
+#include <forward_list>
 #include <list>
+#include <deque>
 
 using namespace ::testing;
 
-TEST_F(TypesAndTraits, Properties)
+TEST(StringAccumulate, AccumulateEqual)
 {
-    static_assert(std::is_same<
-                  map_key_iterator,
-                  pair_iterator<iterator, true>>::value,
-                  "Do not change key iterator type!");
+    std::string result;
 
-    static_assert(std::is_same<
-                  map_value_iterator,
-                  pair_iterator<iterator, false>>::value,
-                  "Do not change value iterator type!");
+    std::vector<std::string> dataToConcatenate {"one",
+                                                "two",
+                                                "three",
+                                                "four"};
+
+    std::copy(dataToConcatenate.begin(), dataToConcatenate.end(),
+              make_accumulator(result));
+
+    std::string expected {std::accumulate(dataToConcatenate.begin(), dataToConcatenate.end(),
+                                           std::string {})};
+
+    EXPECT_EQ(expected, result);
 }
 
-GENERATE_HAS_MEMBER_TYPE(value_type)
-GENERATE_HAS_MEMBER_TYPE(reference)
-GENERATE_HAS_MEMBER_TYPE(pointer)
-GENERATE_HAS_MEMBER_TYPE(iterator_category)
-GENERATE_HAS_MEMBER_TYPE(difference_type)
-
-TEST_F(TypesAndTraits, Traits)
+TEST(StringAccumulate, Forward)
 {
-    static_assert (has_member_type_value_type<map_key_iterator>::value,
-                   "Traits are not provided for key iterator!");
-    static_assert (has_member_type_pointer<map_key_iterator>::value,
-                   "Traits are not provided for key iterator!");
-    static_assert (has_member_type_reference<map_key_iterator>::value,
-                   "Traits are not provided for key iterator!");
-    static_assert (has_member_type_iterator_category<map_key_iterator>::value,
-                   "Traits are not provided for key iterator!");
-    static_assert (has_member_type_difference_type<map_key_iterator>::value,
-                   "Traits are not provided for key iterator!");
+    std::string result;
 
-    static_assert (has_member_type_value_type<map_value_iterator>::value,
-                   "Traits are not provided for value iterator!");
-    static_assert (has_member_type_pointer<map_value_iterator>::value,
-                   "Traits are not provided for value iterator!");
-    static_assert (has_member_type_reference<map_value_iterator>::value,
-                   "Traits are not provided for value iterator!");
-    static_assert (has_member_type_iterator_category<map_value_iterator>::value,
-                   "Traits are not provided for value iterator!");
-    static_assert (has_member_type_difference_type<map_value_iterator>::value,
-                   "Traits are not provided for value iterator!");
+    std::forward_list<std::string> dataToConcatenate {"one",
+                                                      "two",
+                                                      "three",
+                                                      "four"};
+
+    std::copy(dataToConcatenate.begin(), dataToConcatenate.end(),
+              make_accumulator(result));
+
+    std::string expected {std::accumulate(dataToConcatenate.begin(), dataToConcatenate.end(),
+                                           std::string {})};
+
+    EXPECT_EQ(expected, result);
 }
 
-TEST_F(TypesAndTraits, ValueTypeTraits)
+struct move_only
 {
-    static_assert (std::is_same<const key_type,
-                                map_key_iterator::value_type>::value,
-                   "Expected correct key value_type trait");
+    move_only() = default;
+    move_only(move_only&&)
+        : moved {true} {
+    }
+    move_only& operator=(move_only&&) {
+        moved = true;
+        return *this;
+    }
+    bool moved {false};
+};
 
-    static_assert (std::is_same<value_type,
-                                map_value_iterator::value_type>::value,
-                   "Expected correct key type property");
+TEST(Accumulate, RValues)
+{
+    std::vector<move_only> result;
+    const auto internalEntrySize = 10;
+
+    std::vector<std::vector<move_only>> dataToConcatenate;
+    dataToConcatenate.resize(internalEntrySize);
+    for (auto& entry : dataToConcatenate)
+    {
+        entry.resize(internalEntrySize);
+    }
+
+    std::copy(std::make_move_iterator(dataToConcatenate.begin()),
+              std::make_move_iterator(dataToConcatenate.end()),
+              make_accumulator(result));
+
+    EXPECT_EQ(result.size(), internalEntrySize * dataToConcatenate.size());
+    EXPECT_TRUE(std::all_of(result.begin(), result.end(), [](const move_only& object) {
+        return object.moved;
+    }));
 }
 
-TEST_F(Usage, BasicAssign)
+TEST(Accumulate, OtherContainerRvalue)
 {
-    auto keysView = make_keys_view(data);
-    std::vector<std::string> keys {keysView.begin(),
-                                   keysView.end()};
+    std::vector<size_t> result;
+    const auto internalEntrySize = 10;
 
+    std::list<std::deque<size_t>> dataToConcatenate;
+    dataToConcatenate.resize(internalEntrySize);
+    for (auto& entry : dataToConcatenate)
+    {
+        entry.resize(internalEntrySize);
+    }
 
-    auto valuesView = make_values_view(data);
-    std::list<size_t> values {valuesView.begin(),
-                                valuesView.end()};
+    std::copy(std::make_move_iterator(dataToConcatenate.begin()),
+              std::make_move_iterator(dataToConcatenate.end()),
+              make_accumulator(result));
 
-    EXPECT_EQ(data.size(), keys.size());
-    EXPECT_EQ(data.size(), values.size());
+    EXPECT_EQ(result.size(), internalEntrySize * dataToConcatenate.size());
 }
 
-TEST_F(Usage, BasicAlgorithm)
+TEST(Accumulate, OtherContainerLvalue)
 {
-    std::vector<std::string> keys {};
-    std::copy(make_key_iterator(data.begin()),
-              make_key_iterator(data.end()),
-              std::back_inserter(keys));
+    std::vector<size_t> result;
+    const auto internalEntrySize = 10;
 
-    std::list<size_t> values;
-    std::reverse_copy(make_value_iterator(data.begin()),
-                      make_value_iterator(data.end()),
-                      std::back_inserter(values));
+    std::list<std::deque<size_t>> dataToConcatenate;
+    dataToConcatenate.resize(internalEntrySize);
+    for (auto& entry : dataToConcatenate)
+    {
+        entry.resize(internalEntrySize);
+    }
+    const auto constCopy = dataToConcatenate;
+    std::copy(constCopy.cbegin(),
+              constCopy.cend(),
+              make_accumulator(result));
 
-    EXPECT_EQ(data.size(), keys.size());
-    EXPECT_EQ(data.size(), values.size());
-
+    EXPECT_EQ(result.size(), internalEntrySize * constCopy.size());
 }
 
-TEST_F(Usage, BasicRangeFor)
+TEST(Properties, Traits)
 {
-    for (auto& key : make_keys_view(data))
-    {
-        bool canAssign = std::is_assignable<decltype (key), std::string>::value;
-        EXPECT_FALSE(canAssign);
-    }
+    using container = std::vector<int>;
+    using accumulator = accumulator<container>;
+    using category = std::iterator_traits<accumulator>::iterator_category;
 
-    for (const auto& key : make_keys_view(data))
-    {
-        bool canAssign = std::is_assignable<decltype (key), std::string>::value;
-        EXPECT_FALSE(canAssign);
-    }
+    static_assert (std::is_same<category, std::output_iterator_tag>::value,
+                   "Wrong iterator type!");
 
-    for (const auto& value : make_values_view(data))
-    {
-        bool canAssign = std::is_assignable<decltype (value), size_t>::value;
-        EXPECT_FALSE(canAssign);
-    }
+    using reference = std::add_lvalue_reference<accumulator>::type;
+    using return_type_deref = decltype(std::declval<accumulator>().operator*());
+    static_assert (std::is_same<return_type_deref,
+                                reference>::value,
+                   "Wrong return type for dereferencing");
 
-    for (auto& value : make_values_view(data))
-    {
-        bool canAssign = std::is_assignable<decltype (value), size_t>::value;
-        EXPECT_TRUE(canAssign);
-    }
-}
+    using return_type_increment = decltype(std::declval<accumulator>().operator++());
+    static_assert (std::is_same<return_type_increment,
+                                reference>::value,
+                   "Wrong return type for dereferencing");
 
-/** @warning ADVANCED TASK! */
-TEST_F(Usage, DISABLED_Const)
-{
-    const auto constData = data;
-    for (auto& key : make_keys_view(constData))
-    {
-        bool canAssign = std::is_assignable<decltype (key), std::string>::value;
-        EXPECT_FALSE(canAssign);
-    }
-    for (const auto& key : make_keys_view(constData))
-    {
-        bool canAssign = std::is_assignable<decltype (key), std::string>::value;
-        EXPECT_FALSE(canAssign);
-    }
+    static_assert (std::is_trivially_copyable<accumulator>::value,
+                   "Shall support trivial copying");
 
-    EXPECT_TRUE(static_cast<bool>(std::is_same<decltype(make_value_iterator(data.cbegin()))::reference,
-                                               const size_t&>::value));
+    using lvalue_container = std::add_lvalue_reference<container>::type;
+    static_assert (std::is_assignable<accumulator,
+                                      lvalue_container>::value,
+                   "Shall support container assign (lvalue)");
+
+    using rvalue_container = std::add_rvalue_reference<container>::type;
+    static_assert (std::is_assignable<accumulator,
+                                      rvalue_container>::value,
+                   "Shall support container assign (rvalue)");
+
+    using another_container = std::deque<int>;
+    using const_ref_another = std::add_lvalue_reference<std::add_const<another_container>::type>::type;
+    static_assert (std::is_assignable<accumulator,
+                                      const_ref_another>::value,
+                   "Shall support another container assign (lvalue)");
 }
