+++ iterators_task_3/student/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(reverse_words VERSION 1.0.0)
+project(key_value_iterator_HOMEWORK VERSION 1.0.0)
 
 set(sources
     "src/impl.cpp"
@@ -14,19 +14,18 @@ set(BUILDDIR ${ROOT_DIR}/build)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${BUILDDIR}/bin)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${BUILDDIR}/bin)
 set(LESSONS_ROOT ${ROOT_DIR}/../../)
+
 set(TEST ON)
 
-add_library(reverse_words SHARED ${sources})
+add_library(key_value_iterator_HOMEWORK SHARED ${sources})
 
-set_target_properties(reverse_words PROPERTIES
+set_target_properties(key_value_iterator_HOMEWORK PROPERTIES
                         C_STANDARD 11
                         CXX_STANDARD 11
                         PUBLIC_HEADER "${headers}"
 )
 
-
-
-target_include_directories(reverse_words
+target_include_directories(key_value_iterator_HOMEWORK
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
