+++ iterators_task_3/student/src/impl.cpp
@@ -1,2 +1,1 @@
 #include "impl.h"
-
